Add game-over condition.
Right now, this is a primitive approaching using timeouts, which has
some bugs. Specifically, it creates a new timeout each time you drop a
block, without cancelling old ones, so if you drop a bunch of blocks in
a row and then try to restart, it will instantly end your new game.

We'll fix this later using core.async.